requirements.txt
matplotlib==3.7.1
networkx==3.1
numpy==1.24.3
pandas==2.0.1
PyPDF2==3.0.1
python-docx==0.8.11
scikit-learn==1.2.2
seaborn==0.12.2
streamlit==1.22.0
pathlib==1.0.1 
re==2.2.1   
